moved images to the internal folder
