+++ b2b-sample/guardio-insurance-sample-application/components/settingsComponents/homeComponet/newsComponent.js
@@ -19,10 +19,10 @@
 import Image from 'next/image';
 import React from 'react';
 import { Panel } from 'rsuite';
-import image1 from '../../../public/news1.jpeg';
-import image2 from '../../../public/news2.jpeg';
-import image3 from '../../../public/news3.jpeg';
-import image4 from '../../../public/news4.jpeg';
+import image1 from '../../../public/internal/news1.jpeg';
+import image2 from '../../../public/internal/news2.jpeg';
+import image3 from '../../../public/internal/news3.jpeg';
+import image4 from '../../../public/internal/news4.jpeg';
 import { getCurrentDate } from '../../../util/util/common/common';
 
 export default function NewsComponent(props) {
