+++ b2b-sample/guardio-insurance-sample-application/public/internal/home.jpeg
