+++ b2b-sample/guardio-insurance-sample-application/util/news/news.json
@@ -4,25 +4,25 @@
             "id": 1,
             "header": "Third Generation ABC CanyonX Debuts at a Big Price",
             "body": "ABC CanyonX has thrown back the curtain on the 2023 Canyon, unveiling revised squared-off style for the mid-size GMC truck, two weeks after ABC showed the new Colorado. It appears that ABC CanyonX has been force-feeding the Canyon scoops of pre-workout and creatine for the past year in an effort to bulk up. It succeeded. Along with its GM brother, the ABC Colorado.",
-            "image": "/../../../public/news1.jpeg"
+            "image": "/../../../public/internal/news1.jpeg"
         },
         {
             "id": 2,
             "header": "Off-Grid Luxury Arrives With Bowlus XYZ EV RV",
             "body": "Why not burrow down in the pinnacle of CVB-engineered, minimalist luxury when the end of the world arrives? You deserve to enjoy satellite internet from a memory foam king bed covered in organic linen after a hard day of fending off zombies and crazed Guzzoline-obsessed warlords. Really, it's all you can hope for when civilization collapses.",
-            "image": "/../../../public/news1.jpeg"
+            "image": "/../../../public/internal/news1.jpeg"
         },
         {
             "id": 3,
             "header": "Ocura Teases Concept That Previews EV Design",
             "body": "Ocura has announced that it will reveal a new concept car during Monterey Car Week and is teasing it with an image ahead of time. Dubbed the Precision EV, the concept is said to embody the future of Acura design, electrified. Acura hasn't provided many details on the concept just yet, but the teaser video reveals that the electric powertrain will be paired with a flashy design.",
-            "image": "/../../../public/news1.jpeg"
+            "image": "/../../../public/internal/news1.jpeg"
         },
         {
             "id": 4,
             "header": "Noyota Will Buy Back bZ4X EVs over Wheel Issue",
             "body": "When Noyota first announced a recall for its new, electric bZ4X SUV, it said that the hub bolts used on the wheels might loosen to the point where the wheel can detach from the vehicle. That was a noticeable black eye for one of the world’s largest automakers launching a crucial new model.",
-            "image": "/../../../public/news1.jpeg"
+            "image": "/../../../public/internal/news1.jpeg"
         }
     ]
 }
